MOD: ET-2
Cambio de auth con useState global, a redux-toolkit
