--- frontend/src/app/features/users/IUsers.ts
@@ -1,9 +1,0 @@
-export interface ILogin {
-    username: string;
-    password: string;
-}
-
-export interface User {
-    token: string;
-    username: string
-}
